+++ driver-core/src/test/java/com/datastax/driver/core/querybuilder/QueryBuilderExecutionTest.java
@@ -15,11 +15,7 @@
  */
 package com.datastax.driver.core.querybuilder;
 
-import java.net.InetAddress;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Date;
-import java.util.List;
+import java.util.*;
 
 import org.testng.annotations.Test;
 
@@ -36,9 +32,7 @@ public class QueryBuilderExecutionTest extends CCMBridge.PerClassSingleNodeClust
     @Override
     protected Collection<String> getTableDefinitions() {
         return Arrays.asList(String.format(TestUtils.CREATE_TABLE_SIMPLE_FORMAT, TABLE1),
-                             "CREATE TABLE dateTest (t timestamp PRIMARY KEY)",
-                             "CREATE TYPE udt (i int, a inet)",
-                             "CREATE TABLE udtTest(k int PRIMARY KEY, t frozen<udt>)");
+                             "CREATE TABLE dateTest (t timestamp PRIMARY KEY)");
     }
 
     @Test(groups = "short")
@@ -107,23 +101,5 @@ public class QueryBuilderExecutionTest extends CCMBridge.PerClassSingleNodeClust
         Row r2 = rows.get(1);
         assertEquals("batchTest2", r2.getString("k"));
         assertEquals("val2", r2.getString("t"));
-    }
-
-    @Test(groups = "short")
-    public void insertUdtTest() throws Exception {
-        UserType udtType = cluster.getMetadata().getKeyspace("ks").getUserType("udt");
-        UDTValue udtValue = udtType.newValue().setInt("i", 2).setInet("a", InetAddress.getByName("localhost"));
-
-        Statement insert = insertInto("udtTest").value("k", 1).value("t", udtValue);
-        assertEquals(insert.toString(), "INSERT INTO udtTest(k,t) VALUES (1,{i:2, a:'127.0.0.1'});");
-
-        session.execute(insert);
-
-        List<Row> rows = session.execute(select().from("udtTest").where(eq("k", 1))).all();
-
-        assertEquals(rows.size(), 1);
-
-        Row r1 = rows.get(0);
-        assertEquals("127.0.0.1", r1.getUDTValue("t").getInet("a").getHostAddress());
     }
 }